// src/app/models/admin/dangkyhoatdongkhoangsan/dkthamdokhoangsan.model.ts
export class OutputDkThamDoKhoangSanModel {
  iddangkythamdo: string;
  diadiem: string;
  dientichthamdo: number;
  chieusauthamdotu: number;
  chieusauthamdoden: number;
  thoihanthamdo: number;
  mucdichsudungkhoangsan: string;
  dangkhoangsan: number;
  donvidientich: string;
  donvithoihan: string;
  donvichieusau: string;
  hequychieu: string;
  idhoso: string;
  siteid: string;
}

export class InputDkThamDoKhoangSanModel {
  iddangkythamdo: string;
  diadiem: string;
  dientichthamdo: number;
  chieusauthamdotu: number;
  chieusauthamdoden: number;
  thoihanthamdo: number;
  mucdichsudungkhoangsan: string;
  dangkhoangsan: number;
  donvidientich: string;
  donvithoihan: string;
  donvichieusau: string;
  hequychieu: string;
  idhoso: string;
  siteid: string;
}